requirements.txt
einops==0.4.1
fire==0.4.0
gin_config==0.5.0
numpy==1.22.0
pandas==1.4.2
scikit_learn==1.0.2
scipy==1.7.3
torch==1.11.0
tqdm==4.62.3
